Agathe Bauer

+++ irr/include/fast_atof.h
@@ -307,7 +307,7 @@ inline const char *fast_atof_move(const char *in, f32 &result)
 		if (numDecimals < IRR_ATOF_TABLE_SIZE) {
 			value += decimal * fast_atof_table[numDecimals];
 		} else {
-			value += decimal * std::powf(10.f, -(float)numDecimals);
+			value += decimal * std::pow(10.f, -(float)numDecimals);
 		}
 		in = afterDecimal;
 	}
@@ -318,7 +318,7 @@ inline const char *fast_atof_move(const char *in, f32 &result)
 		// strtol10() will deal with both + and - signs,
 		// but calculate as f32 to prevent overflow at FLT_MAX
 		// Using pow with float cast instead of powf as otherwise accuracy decreases.
-		value *= std::powf(10.f, (f32)strtol10(in, &in));
+		value *= std::pow(10.f, (f32)strtol10(in, &in));
 	}
 
 	result = negative ? -value : value;
